requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
redis==4.2.2
wrapt==1.14.0
